Task 4. Updates - Fixed Exception's files
